browser-inference/cnn/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
datasets>=2.14.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=9.5.0
onnx>=1.15.0
onnxruntime>=1.16.0
onnxscript>=0.1.0
